# documentation/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Golden Lift API",
    "version": "0.4.0",
    "description": "Implemented B1-B4 and Dynamic Catalog Core: anonymous browsing/product projection, live staff Identity, complete category administration, dynamic product types/attributes/groups/units and guarded schema/product workflows. Dynamic operations require reviewed Catalog v1.2 cutover. Media delivery, full product/technical editing UX and Inquiry workflows remain later milestones."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/categories": {
      "get": {
        "operationId": "listCategories",
        "parameters": [
          {
            "name": "parentId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "locale",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Locale"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 1024
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryPage"
                }
              }
            }
          },
          "400": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/categories/{id}": {
      "get": {
        "operationId": "getCategory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "locale",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Locale"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/categories/{id}": {
      "patch": {
        "operationId": "editCategory",
        "description": "ADMIN only. Stale expectedVersion returns 409.",
        "security": [
          {
            "staffSession": []
          },
          {
            "staffLocalSession": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "description": "Must exactly match an approved staff origin."
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ActionToken"
            },
            "description": "Read csrfToken from login/current-session JSON; it belongs to the current cookie."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryEditInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "getAdminCategory",
        "description": "ADMIN detail including stored translations, missing locales, parent/cover IDs, exact versions/order and active eligibility. Deleted ancestors exclude the row.",
        "security": [
          {
            "staffSession": []
          },
          {
            "staffLocalSession": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "locale",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Locale"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCategory"
                }
              }
            }
          },
          "400": {
            "description": "Malformed input or cursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, revoked, disabled or deleted session/account.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Requires ADMIN; SUPER_ADMIN is excluded. Unsafe Origin/CSRF is denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Selected category, parent or route is unavailable/deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Stale row, list, path or preview precondition. Reload before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "JSON body exceeds 64 KiB.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Invalid destination, membership, cover, hierarchy or ordering-gap recovery beyond the documented bound.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Identity/Catalog is unavailable or the bounded transaction retry budget is exhausted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteCategoryBranch",
        "description": "ADMIN confirmed atomic soft deletion using the reviewed routine and owner hooks. Revalidates all owned branch state and referenced asset/sheet versions. Descendant, product or attachment changes invalidate preview independently of root version. One bounded versioned outbox event. Fresh Catalog public/admin reads lose branch access; B4 does not revoke Media delivery, signed URLs or downloaded files. Repeated request is 404.",
        "security": [
          {
            "staffSession": []
          },
          {
            "staffLocalSession": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "description": "Configured staff origin."
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Session-bound token from login/session."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteBranchInput"
              },
              "example": {
                "confirm": true,
                "expectedVersion": "1",
                "previewPrecondition": "b1-bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchDeletionResult"
                }
              }
            }
          },
          "400": {
            "description": "Malformed input or cursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, revoked, disabled or deleted session/account.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Requires ADMIN; SUPER_ADMIN is excluded. Unsafe Origin/CSRF is denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Selected category, parent or route is unavailable/deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Stale row, list, path or preview precondition. Reload before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "JSON body exceeds 64 KiB.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Invalid destination, membership, cover, hierarchy or ordering-gap recovery beyond the documented bound.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Identity/Catalog is unavailable or the bounded transaction retry budget is exhausted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Locale": {
        "type": "string",
        "enum": ["ar", "en", "ckb"],
        "default": "ar"
      },
      "Version": {
        "type": "string",
        "pattern": "^[1-9][0-9]{0,18}$",
        "description": "Positive signed bigint represented without JavaScript precision loss."
      },
...
      "ApiError": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message", "requestId"],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              }
            }
          }
        }
      },
      "Category": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "parentId",
          "name",
          "description",
          "slug",
          "locale",
          "resolvedNameLocale",
          "sortOrder",
          "version"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "parentId": {
            "type": ["string", "null"],
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": ["string", "null"]
          },
          "slug": {
            "type": ["string", "null"]
          },
          "locale": {
            "$ref": "#/components/schemas/Locale"
          },
          "resolvedNameLocale": {
            "$ref": "#/components/schemas/Locale"
          },
          "sortOrder": {
            "type": "string",
            "pattern": "^-?[0-9]+$"
          },
          "version": {
            "$ref": "#/components/schemas/Version"
          }
        }
      },
      "CategoryPage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["items", "nextCursor"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Category"
            }
          },
          "nextCursor": {
            "type": ["string", "null"]
          }
        }
      },
...
      "ActionToken": {
        "type": "string",
        "pattern": "^[A-Za-z0-9_-]{43}$"
      },
...
      "CategoryTranslation": {
        "type": "object",
        "additionalProperties": false,
        "required": ["locale", "name"],
        "properties": {
          "locale": {
            "$ref": "#/components/schemas/Locale"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 300
          },
          "description": {
            "type": ["string", "null"]
          },
          "slug": {
            "type": ["string", "null"]
          }
        }
      },
...
      "CategoryEditInput": {
        "type": "object",
        "additionalProperties": false,
        "required": ["expectedVersion", "translations"],
        "properties": {
          "expectedVersion": {
            "$ref": "#/components/schemas/Version"
          },
          "translations": {
            "type": "array",
            "minItems": 1,
            "maxItems": 3,
            "items": {
              "$ref": "#/components/schemas/CategoryTranslation"
            },
            "description": "Unique locales; an Arabic name is mandatory."
          },
          "coverAssetId": {
            "type": ["string", "null"],
            "format": "uuid"
          }
        },
        "description": "Supply current expectedVersion and unique saved translations, including valid Arabic. Optional coverAssetId must refer to an already verified active image registration; null clears a cover, omission preserves it. Parent/order changes are rejected on generic PATCH."
      },
      "CategoryListRevision": {
        "type": "string",
        "pattern": "^l1-[a-f0-9]{64}$",
        "description": "Opaque scope-bound state precondition. Re-read after 409; never an authorization credential."
      },
...
      "BranchPreviewPrecondition": {
        "type": "string",
        "pattern": "^b1-[a-f0-9]{64}$",
        "description": "Opaque scope-bound state precondition. Re-read after 409; never an authorization credential."
      },
      "StoredCategoryTranslation": {
        "type": "object",
        "additionalProperties": false,
        "required": ["locale", "name", "description", "slug", "version"],
        "properties": {
          "locale": {
            "$ref": "#/components/schemas/Locale"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 300
          },
          "description": {
            "type": ["string", "null"]
          },
          "slug": {
            "type": ["string", "null"]
          },
          "version": {
            "$ref": "#/components/schemas/Version"
          }
        }
      },
      "AdminCategory": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "parentId",
          "name",
          "description",
          "slug",
          "locale",
          "resolvedNameLocale",
          "sortOrder",
          "version",
          "coverAssetId",
          "translations",
          "missingTranslationLocales",
          "activeChildCount",
          "activeProductCount",
          "canAddChildren",
          "canAddProducts"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "parentId": {
            "type": ["string", "null"],
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": ["string", "null"]
          },
          "slug": {
            "type": ["string", "null"]
          },
          "locale": {
            "$ref": "#/components/schemas/Locale"
          },
          "resolvedNameLocale": {
            "$ref": "#/components/schemas/Locale"
          },
          "sortOrder": {
            "type": "string",
            "pattern": "^-?[0-9]+$"
          },
          "version": {
            "$ref": "#/components/schemas/Version"
          },
          "coverAssetId": {
            "type": ["string", "null"],
            "format": "uuid"
          },
          "translations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StoredCategoryTranslation"
            }
          },
          "missingTranslationLocales": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Locale"
            }
          },
          "activeChildCount": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$"
          },
          "activeProductCount": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$"
          },
          "canAddChildren": {
            "type": "boolean"
          },
          "canAddProducts": {
            "type": "boolean"
          }
        },
        "description": "ADMIN only. Display fields fall back to Arabic per field; translations contain actual saved values. Missing locales are explicit. Eligibility derives from active child/product counts; there is no stored mode."
      },
...
      "BranchDeletionImpact": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "totalCategoryCount",
          "descendantCategoryCount",
          "productCount",
          "categoryTranslationCount",
          "categorySpecificationCount",
          "categoryCoverCount",
          "categoryTechnicalLinkCount",
          "productTranslationCount",
          "productCodeReservationCount",
          "productMediaCount",
          "productMediaTranslationCount",
          "productSpecificationValueCount",
          "productSpecificationTextCount",
          "productSpecificationChoiceCount",
          "productTechnicalLinkCount",
          "productTechnicalConfigurationCount"
        ],
        "properties": {
          "totalCategoryCount": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$"
          },
          "descendantCategoryCount": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$"
          },
          "productCount": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$"
          },
          "categoryTranslationCount": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$"
          },
          "categorySpecificationCount": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$"
          },
          "categoryCoverCount": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$"
          },
          "categoryTechnicalLinkCount": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$"
          },
          "productTranslationCount": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$"
          },
          "productCodeReservationCount": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$"
          },
          "productMediaCount": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$"
          },
          "productMediaTranslationCount": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$"
          },
          "productSpecificationValueCount": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$"
          },
          "productSpecificationTextCount": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$"
          },
          "productSpecificationChoiceCount": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$"
          },
          "productTechnicalLinkCount": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$"
          },
          "productTechnicalConfigurationCount": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$"
          }
        },
        "description": "Active records only. Total includes selected root; descendants exclude it. Category covers stop being publicly referenced; registrations remain. Owned technical links/selections are deleted, shared sheet contents and source evidence remain. Counts use exact decimal strings."
      },
...
      "DeleteBranchInput": {
        "type": "object",
        "additionalProperties": false,
        "required": ["confirm", "expectedVersion", "previewPrecondition"],
        "properties": {
          "confirm": {
            "const": true
          },
          "expectedVersion": {
            "$ref": "#/components/schemas/Version"
          },
          "previewPrecondition": {
            "$ref": "#/components/schemas/BranchPreviewPrecondition"
          }
        },
        "description": "Use the selected root version and scope-bound token from deletion-preview. Explicit boolean true required. Recompute full branch state in the deleting transaction; any relevant owned-record/membership change produces 409, even if the root version is unchanged. No restore; repeated deletion is 404."
      },
      "BranchDeletionResult": {
        "type": "object",
        "additionalProperties": false,
        "required": ["categoryId", "version", "impact", "sourceListRevision"],
        "properties": {
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "version": {
            "$ref": "#/components/schemas/Version"
          },
          "impact": {
            "$ref": "#/components/schemas/BranchDeletionImpact"
          },
          "sourceListRevision": {
            "$ref": "#/components/schemas/CategoryListRevision"
          }
        }
      },
...
    },
    "securitySchemes": {
      "staffSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-gl_staff",
        "description": "Production HttpOnly, Secure, SameSite=Strict cookie. All protected mutations also require the session-bound X-CSRF-Token and approved Origin."
      },
      "staffLocalSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "gl_staff",
        "description": "Development-only cookie, never used in production."
      }
    }
  }
}